# GPKG: add ST_EnvIntersects() for faster spatial filtering when there is no spatial index