@jratike80 2023-04-20
Is the somehow ungrammatical ST_EnvelopesIntersects alias there because SpatiaLite has it also?